Fourth Commit
Need to fix write methods, null case for costing data, and update excel colors.

+++ JobCosting/StockJob.cs
@@ -17,7 +17,7 @@ namespace JobCosting
         /// </summary>
         public override void setAmountActualCost()
         {            
-            amountActualCost = amountActualCost + (double)productCost * (double)orderQuantity;          
+            amountActualCost = amountActualCost + productCost * orderQuantity;          
         }
     }
 }
